requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
Django==4.0.1
djutils==1.0.15
elastic-transport==8.1.2
elasticsearch==8.1.2
entrypoints==0.3
executing==0.8.2
fake-useragent==0.1.11
filelock==3.6.0
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
lxml==4.8.0
matplotlib-inline==0.1.3
MechanicalSoup==1.1.0
multidict==6.0.2
mypy-extensions==0.4.3
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
parse==1.19.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.4.1
playsound==1.3.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
pure-eval==0.2.1
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pygame==2.1.2
Pygments==2.11.2
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
python-vlc==3.0.12118
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
schedule==1.1.0
simpleaudio==1.0.4
six==1.16.0
snscrape==0.4.3.20220107.dev45+ged3ea94
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.3
yarl==1.7.2
zipp==3.8.0
